// src/app/email/input/smtp.interface.ts
export interface mailConfig {
  host: string;
  port: number;
  secure: boolean;
  auth: {
    user: string;
    pass: string;
  };
}

export interface mailAttachment {
  filename?: string;
  content?: string; // Can be base64 string
  path?: string; // Can be a URL or local file path
  contentType?: string;
  cid?: string; // Content ID for inline images
}

export interface mailInput {
  from: string;
  to: string[];
  subject: string;
  html?: string;
  text?: string;
  attachments?: mailAttachment[];
}

export interface SendMailInput {
  config: mailConfig;
  mail: mailInput;
}

export interface SendMailOutput {
  success: boolean;
  message: string;
  data: {
    accepted: string[];
    rejected: string[];
    response: string[];
  };
}